+++ divrec/train/__init__.py
@@ -1,4 +1,5 @@
 from .base_trainer import Trainer
+from .bpr_trainer import BPRTrainer
 
 
-__all__ = ["Trainer"]
+__all__ = ["Trainer", "BPRTrainer"]
