+++ divrec_experiments/experiments/matrix_factorization_ml100k/src/__init__.py
@@ -1,0 +1,10 @@
+from .load_stage import load_dataset
+from .test_stage import test_model
+from .train_stage import train_model
+
+
+__all__ = [
+    "load_dataset",
+    "test_model",
+    "train_model",
+]
